farmer: Refactor `verify_pieces_at_blockchain` method.
